Move the odoo connector into the admin directory
* It should only be used by the admin pages.
