+++ admin/main.php
@@ -1,5 +1,6 @@
 <?php
 
+require_once 'odoo_connector/odoo_connector.php';
 require_once 'database_connection/main.php';
 require_once 'api/main.php';
 require_once 'menu_items.php';
