+++ tests/php/admin/odoo_connector/odoo_connector/OdooConnOdooConnectorTestBase.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace odoo_conn\tests\odoo_connector\odoo_connector;
+namespace php\admin\odoo_connector\odoo_connector;
 
-require_once(__DIR__ . "/OdooConnOdooConnectorTestBase.php");
+require_once __DIR__ . "/OdooConnOdooConnectorTestBase.php";
 
-use \Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
+use Mockery;
+use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
 use PHPUnit\Framework\TestCase;
 use PhpXmlRpc\Value;
 
@@ -15,9 +16,9 @@ class OdooConnOdooConnectorTestBase extends TestCase
 
     public function setUp(): void
     {
-        $this->odoo_connector = \Mockery::mock(
-            "\odoo_conn\odoo_connector\odoo_connector\OdooConnOdooConnector[create_client, create_request, create_value]",
-            array("username", "api_key", "database", "test_url")
+        $this->odoo_connector = Mockery::mock(
+            "\odoo_conn\admin\odoo_connector\OdooConnOdooConnector[create_client, create_request, create_value]",
+            ["username", "api_key", "database", "test_url"]
         )->makePartial();
 
         $this->odoo_connector->shouldReceive("create_value")
@@ -28,14 +29,14 @@ class OdooConnOdooConnectorTestBase extends TestCase
                 return $value;
             });
 
-        $this->version_request = \Mockery::mock();
+        $this->version_request = Mockery::mock();
         $this->odoo_connector->shouldReceive("create_request")->with(
             "version", []
         )->andReturn($this->version_request);
-        $this->version_response = \Mockery::mock();
+        $this->version_response = Mockery::mock();
         $this->version_response->shouldReceive("value")->with()->andReturn("14.0");
 
-        $this->authentication_request = \Mockery::mock();
+        $this->authentication_request = Mockery::mock();
         $this->odoo_connector->shouldReceive("create_request")->with(
             "authenticate",
             [
@@ -46,7 +47,7 @@ class OdooConnOdooConnectorTestBase extends TestCase
             ]
         )->andReturn($this->authentication_request);
 
-        $this->common_client = \Mockery::mock();
+        $this->common_client = Mockery::mock();
         $this->odoo_connector->shouldReceive("create_client")->with(
             "test_url/xmlrpc/2/common"
         )->once()->andReturn(
